Change limiter & validation
