+++ middlewares/limiter.js
@@ -1,0 +1,8 @@
+const rateLimit = require('express-rate-limit');
+
+const limiter = rateLimit({
+  windowMs: 15 * 60 * 1000,
+  max: 100,
+});
+
+module.exports = limiter;
